Switch kernels to work with uchar rather than floats

+++ include/device.h
@@ -50,7 +50,7 @@ public:
 
 	void CopyBuffer (const cl::Buffer& src, cl::Buffer& dest, size_t size);
 
-	void Mask(const cl::Buffer& d_out, cl::Buffer& d_in, cl::Buffer& d_mask, float mask_value, size_t m, size_t n, size_t local_size_m, size_t local_size_n);
+	void Mask(const cl::Buffer& d_out, cl::Buffer& d_in, cl::Buffer& d_mask, uint8_t mask_value, size_t m, size_t n, size_t local_size_m, size_t local_size_n);
 
 	void Upcast(const cl::Buffer& d_out, cl::Buffer& d_in, size_t len, size_t local_size);
 
@@ -65,6 +65,8 @@ public:
 	void FlagRows(const cl::Buffer& mask, float row_sum_threshold, size_t m, size_t n, size_t local_size);
 
 	float Reduce(const cl::Buffer d_in, size_t drop_out, size_t len, size_t local_size);
+
+	//void Grubb(const cl::Buffer data, size_t len, size_t work_size, local_size);
 
 };
